requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pymysql
jinja2
yfinance
pandas
numpy
requests
python-dotenv
python-dateutil
akshare
